search: Remove unused code for search pills.
