+++ frontend_tests/node_tests/search.js
@@ -5,40 +5,27 @@ const {strict: assert} = require("assert");
 const {mock_esm, set_global, zrequire} = require("../zjsunit/namespace");
 const {run_test} = require("../zjsunit/test");
 const $ = require("../zjsunit/zjquery");
-const {page_params} = require("../zjsunit/zpage_params");
 
 const noop = () => {};
 
 const narrow = mock_esm("../../static/js/narrow");
-const narrow_state = mock_esm("../../static/js/narrow_state", {
-    filter: () => false,
-});
+const narrow_state = mock_esm("../../static/js/narrow_state");
 const search_suggestion = mock_esm("../../static/js/search_suggestion");
-
-mock_esm("../../static/js/search_pill_widget", {
-    widget: {
-        getByID: () => true,
-    },
-});
 mock_esm("../../static/js/ui_util", {
     change_tab_to: noop,
-    place_caret_at_end: noop,
+});
+
+const Filter = {};
+
+mock_esm("../../static/js/filter", {
+    Filter,
 });
 
 set_global("setTimeout", (func) => func());
 
 const search = zrequire("search");
-const search_pill = zrequire("search_pill");
-const {Filter} = zrequire("../js/filter");
 
-function test(label, f) {
-    run_test(label, ({override}) => {
-        page_params.search_pills_enabled = true;
-        f({override});
-    });
-}
-
-test("clear_search_form", () => {
+run_test("clear_search_form", () => {
     $("#search_query").val("noise");
     $("#search_query").trigger("focus");
     $(".search_button").prop("disabled", false);
@@ -50,7 +37,7 @@ test("clear_search_form", () => {
     assert.equal($(".search_button").prop("disabled"), true);
 });
 
-test("update_button_visibility", () => {
+run_test("update_button_visibility", () => {
     const $search_query = $("#search_query");
     const $search_button = $(".search_button");
 
@@ -83,24 +70,18 @@ test("update_button_visibility", () => {
     assert.ok(!$search_button.prop("disabled"));
 });
 
-test("initialize", () => {
+run_test("initialize", () => {
     const $search_query_box = $("#search_query");
     const $searchbox_form = $("#searchbox_form");
     const $search_button = $(".search_button");
-    const $searchbox = $("#searchbox");
 
-    $search_query_box[0] = "stub";
-
-    search_pill.get_search_string_for_current_filter = () => "is:starred";
-
-    search_suggestion.max_num_of_search_results = 99;
+    search_suggestion.max_num_of_search_results = 999;
     $search_query_box.typeahead = (opts) => {
         assert.equal(opts.fixed, true);
-        assert.equal(opts.items, 99);
+        assert.equal(opts.items, 999);
         assert.equal(opts.naturalSearch, true);
         assert.equal(opts.helpOnEmptyStrings, true);
         assert.equal(opts.matcher(), true);
-        opts.on_move();
 
         {
             const search_suggestions = {
@@ -143,17 +124,12 @@ test("initialize", () => {
         {
             let operators;
             let is_blurred;
-            let is_append_search_string_called;
             $search_query_box.on("blur", () => {
                 is_blurred = true;
             });
-            search_pill.append_search_string = () => {
-                is_append_search_string_called = true;
-            };
             /* Test updater */
             const _setup = (search_box_val) => {
                 is_blurred = false;
-                is_append_search_string_called = false;
                 $search_query_box.val(search_box_val);
                 Filter.parse = (search_string) => {
                     assert.equal(search_string, search_box_val);
@@ -163,7 +139,6 @@ test("initialize", () => {
                     assert.deepEqual(raw_operators, operators);
                     assert.deepEqual(options, {trigger: "search"});
                 };
-                search_pill.get_search_string_for_current_filter = () => search_box_val;
             };
 
             operators = [
@@ -174,10 +149,8 @@ test("initialize", () => {
                 },
             ];
             _setup("ver");
-
             assert.equal(opts.updater("ver"), "ver");
-            assert.ok(!is_blurred);
-            assert.ok(is_append_search_string_called);
+            assert.ok(is_blurred);
 
             operators = [
                 {
@@ -187,17 +160,13 @@ test("initialize", () => {
                 },
             ];
             _setup("stream:Verona");
-
             assert.equal(opts.updater("stream:Verona"), "stream:Verona");
-            assert.ok(!is_blurred);
-            assert.ok(is_append_search_string_called);
+            assert.ok(is_blurred);
 
             search.__Rewire__("is_using_input_method", true);
             _setup("stream:Verona");
-
             assert.equal(opts.updater("stream:Verona"), "stream:Verona");
             assert.ok(!is_blurred);
-            assert.ok(is_append_search_string_called);
 
             $search_query_box.off("blur");
         }
@@ -205,23 +174,14 @@ test("initialize", () => {
 
     search.initialize();
 
-    const $search_pill_stub = $.create(".pill");
-    $search_pill_stub.closest = () => ({data: noop});
-    const stub_event = {
-        // FIXME: event.relatedTarget should not be a jQuery object
-        relatedTarget: $search_pill_stub,
-    };
+    $search_button.prop("disabled", true);
+    $search_query_box.trigger("focus");
+    assert.ok(!$search_button.prop("disabled"));
+
     $search_query_box.val("test string");
     narrow_state.search_string = () => "ver";
-    $search_query_box.trigger(new $.Event("blur", stub_event));
+    $search_query_box.trigger("blur");
     assert.equal($search_query_box.val(), "test string");
-
-    let css_args;
-    $searchbox.css = (args) => {
-        css_args = args;
-    };
-    $searchbox.trigger("focusout");
-    assert.deepEqual(css_args, {"box-shadow": "unset"});
 
     search.__Rewire__("is_using_input_method", false);
     $searchbox_form.trigger("compositionend");
@@ -231,7 +191,7 @@ test("initialize", () => {
     let default_prevented = false;
     let ev = {
         type: "keydown",
-        key: "a",
+        which: 15,
         preventDefault() {
             default_prevented = true;
         },
@@ -249,6 +209,9 @@ test("initialize", () => {
     assert.equal(keydown(ev), undefined);
     assert.ok(default_prevented);
 
+    ev = {
+        type: "keyup",
+    };
     let operators;
     let is_blurred;
     narrow_state.active = () => false;
@@ -269,7 +232,6 @@ test("initialize", () => {
             assert.deepEqual(raw_operators, operators);
             assert.deepEqual(options, {trigger: "search"});
         };
-        search_pill.get_search_string_for_current_filter = () => search_box_val;
     };
 
     operators = [
@@ -281,10 +243,7 @@ test("initialize", () => {
     ];
     _setup("");
 
-    ev = {
-        type: "keyup",
-        which: 15,
-    };
+    ev.key = "a";
     $search_query_box.is = () => false;
     $searchbox_form.trigger(ev);
 
@@ -316,24 +275,16 @@ test("initialize", () => {
     $searchbox_form.trigger(ev);
     assert.ok(is_blurred);
     assert.ok(!$search_button.prop("disabled"));
-
-    $search_button.prop("disabled", true);
-    $search_query_box.trigger("focus");
-    assert.ok(!$search_button.prop("disabled"));
 });
 
-test("initiate_search", () => {
+run_test("initiate_search", () => {
     // open typeahead and select text when navbar is open
     // this implicitly expects the code to used the chained
     // function calls, which is something to keep in mind if
     // this test ever fails unexpectedly.
+    narrow_state.filter = () => ({is_search: () => true});
     let typeahead_forced_open = false;
     let is_searchbox_text_selected = false;
-    let is_searchbox_focused = false;
-    $("#search_query").off("focus");
-    $("#search_query").on("focus", () => {
-        is_searchbox_focused = true;
-    });
     $("#search_query").typeahead = (lookup) => {
         if (lookup === "lookup") {
             typeahead_forced_open = true;
@@ -344,17 +295,23 @@ test("initiate_search", () => {
         is_searchbox_text_selected = true;
     });
 
-    $("#search_query")[0] = "stub";
+    let searchbox_css_args;
 
-    const $searchbox = $("#searchbox");
-    let css_args;
-    $searchbox.css = (args) => {
-        css_args = args;
+    $("#searchbox").css = (args) => {
+        searchbox_css_args = args;
     };
 
     search.initiate_search();
     assert.ok(typeahead_forced_open);
     assert.ok(is_searchbox_text_selected);
-    assert.ok(is_searchbox_focused);
-    assert.deepEqual(css_args, {"box-shadow": "inset 0px 0px 0px 2px hsl(204, 20%, 74%)"});
+    assert.equal($("#search_query").val(), "ver");
+
+    assert.deepEqual(searchbox_css_args, {
+        "box-shadow": "inset 0px 0px 0px 2px hsl(204, 20%, 74%)",
+    });
+
+    // test that we append space for user convenience
+    narrow_state.filter = () => ({is_search: () => false});
+    search.initiate_search();
+    assert.equal($("#search_query").val(), "ver ");
 });
