+++ static/js/search.js
@@ -4,9 +4,6 @@ import {Filter} from "./filter";
 import * as message_view_header from "./message_view_header";
 import * as narrow from "./narrow";
 import * as narrow_state from "./narrow_state";
-import {page_params} from "./page_params";
-import * as search_pill from "./search_pill";
-import * as search_pill_widget from "./search_pill_widget";
 import * as search_suggestion from "./search_suggestion";
 import * as ui_util from "./ui_util";
 
@@ -23,18 +20,7 @@ export function narrow_or_search_for_term(search_string) {
     }
     ui_util.change_tab_to("#message_feed_container");
 
-    let operators;
-    if (page_params.search_pills_enabled) {
-        // We have to take care to append the new pill before calling this
-        // function, so that the base_query includes the suggestion selected
-        // along with query corresponding to the existing pills.
-        const base_query = search_pill.get_search_string_for_current_filter(
-            search_pill_widget.widget,
-        );
-        operators = Filter.parse(base_query);
-    } else {
-        operators = Filter.parse(search_string);
-    }
+    const operators = Filter.parse(search_string);
     narrow.activate(operators, {trigger: "search"});
 
     // It's sort of annoying that this is not in a position to
@@ -43,9 +29,7 @@ export function narrow_or_search_for_term(search_string) {
 
     // Narrowing will have already put some operators in the search box,
     // so leave the current text in.
-    if (!page_params.search_pills_enabled) {
-        $search_query_box.trigger("blur");
-    }
+    $search_query_box.trigger("blur");
     return $search_query_box.val();
 }
 
@@ -66,7 +50,6 @@ export function update_button_visibility() {
 export function initialize() {
     const $search_query_box = $("#search_query");
     const $searchbox_form = $("#searchbox_form");
-    const $searchbox = $("#searchbox");
 
     // Data storage for the typeahead.
     // This maps a search string to an object with a "description" field.
@@ -77,12 +60,7 @@ export function initialize() {
 
     $search_query_box.typeahead({
         source(query) {
-            let base_query = "";
-            if (page_params.search_pills_enabled) {
-                base_query = search_pill.get_search_string_for_current_filter(
-                    search_pill_widget.widget,
-                );
-            }
+            const base_query = "";
             const suggestions = search_suggestion.get_suggestions(base_query, query);
             // Update our global search_map hash
             search_map = suggestions.lookup_table;
@@ -100,23 +78,12 @@ export function initialize() {
             return true;
         },
         updater(search_string) {
-            if (page_params.search_pills_enabled) {
-                search_pill.append_search_string(search_string, search_pill_widget.widget);
-                return $search_query_box.val();
-            }
             return narrow_or_search_for_term(search_string);
         },
         sorter(items) {
             return items;
         },
-        stopAdvance: page_params.search_pills_enabled,
         advanceKeyCodes: [8],
-
-        on_move() {
-            if (page_params.search_pills_enabled) {
-                ui_util.place_caret_at_end($search_query_box[0]);
-            }
-        },
         // Use our custom typeahead `on_escape` hook to exit
         // the search bar as soon as the user hits Esc.
         on_escape: message_view_header.exit_search,
@@ -167,7 +134,7 @@ export function initialize() {
     // more work to re-order everything and make them private.
 
     $search_query_box.on("focus", focus_search);
-    $search_query_box.on("blur", (e) => {
+    $search_query_box.on("blur", () => {
         // The search query box is a visual cue as to
         // whether search or narrowing is active.  If
         // the user blurs the search box, then we should
@@ -183,32 +150,10 @@ export function initialize() {
         // short enough that the user won't notice (though
         // really it would be OK if they did).
 
-        if (page_params.search_pills_enabled) {
-            const pill_id = $(e.relatedTarget).closest(".pill").data("id");
-            const search_pill = search_pill_widget.widget.getByID(pill_id);
-            if (search_pill) {
-                // The searchbox loses focus while the search
-                // pill element gains focus.
-                // We do not consider the searchbox to actually
-                // lose focus when a pill inside it gets selected
-                // or deleted by a click.
-                return;
-            }
-        }
         setTimeout(() => {
             update_button_visibility();
         }, 100);
     });
-
-    if (page_params.search_pills_enabled) {
-        // Uses jquery instead of pure css as the `:focus` event occurs on `#search_query`,
-        // while we want to add box-shadow to `#searchbox`. This could have been done
-        // with `:focus-within` CSS selector, but it is not supported in IE or Opera.
-        $searchbox.on("focusout", () => {
-            message_view_header.close_search_bar_and_open_narrow_description();
-            $searchbox.css({"box-shadow": "unset"});
-        });
-    }
 }
 
 export function focus_search() {
@@ -220,10 +165,6 @@ export function initiate_search() {
     message_view_header.open_search_bar_and_close_narrow_description();
     $("#searchbox").css({"box-shadow": "inset 0px 0px 0px 2px hsl(204, 20%, 74%)"});
     $("#search_query").typeahead("lookup").trigger("select");
-    if (page_params.search_pills_enabled) {
-        $("#search_query").trigger("focus");
-        ui_util.place_caret_at_end($("#search_query")[0]);
-    }
 }
 
 export function clear_search_form() {
