+++ static/js/ui_init.js
@@ -69,7 +69,6 @@ import * as resize from "./resize";
 import * as rows from "./rows";
 import * as scroll_bar from "./scroll_bar";
 import * as search from "./search";
-import * as search_pill_widget from "./search_pill_widget";
 import * as sent_messages from "./sent_messages";
 import * as server_events from "./server_events";
 import * as settings from "./settings";
@@ -191,7 +190,6 @@ function initialize_right_sidebar() {
 function initialize_navbar() {
     const rendered_navbar = render_navbar({
         embedded: page_params.narrow_stream !== undefined,
-        search_pills_enabled: page_params.search_pills_enabled,
     });
 
     $("#navbar-container").html(rendered_navbar);
@@ -630,7 +628,6 @@ export function initialize_everything() {
     server_events.initialize();
     user_status.initialize(user_status_params);
     compose_pm_pill.initialize();
-    search_pill_widget.initialize();
     reload.initialize();
     user_groups.initialize(user_groups_params);
     unread.initialize();
