+++ static/templates/navbar.hbs
@@ -15,21 +15,6 @@
                 <div class="top-navbar-border top-navbar-container">
                     <div id="message_view_header" class="notdisplayed">
                     </div>
-                    {{#if search_pills_enabled }}
-                    <div id="searchbox">
-                        <form id="searchbox_form" class="form-search navbar-search">
-                            <div id="search_arrows" class="pill-container input-append">
-                                <span class="search_icon search_open" ><i class="fa fa-search" aria-hidden="true"></i></span>
-                                <div contenteditable="true" class="input search-query input-block-level" id="search_query" type="text" placeholder="{{t 'Search' }}"
-                                  autocomplete="off" aria-label="{{t 'Search' }}" title="{{t 'Search' }} (/)">
-                                </div>
-                                <button class="btn search_button" type="button" id="search_exit" aria-label="{{t 'Exit search' }}">
-                                    <i class="fa fa-remove" aria-hidden="true"></i>
-                                </button>
-                            </div>
-                        </form>
-                    </div>
-                    {{else}}
                     <div id="searchbox_legacy">
                         <form id="searchbox_form" class="form-search navbar-search">
                             <div id="search_arrows" class="input-append">
@@ -40,7 +25,6 @@
                             </div>
                         </form>
                     </div>
-                    {{/if}}
                 </div>
             </div>
         </div>
